Switch to golangci-lint
