+++ cmd/internal/cli/build.go
@@ -127,27 +127,6 @@ func checkBuildTarget(path string, update bool) bool {
 	return true
 }
 
-func checkSections() error {
-	var all, none bool
-	for _, section := range sections {
-		if section == "none" {
-			none = true
-		}
-		if section == "all" {
-			all = true
-		}
-	}
-
-	if all && len(sections) > 1 {
-		return fmt.Errorf("Section specification error: Cannot have all and any other option")
-	}
-	if none && len(sections) > 1 {
-		return fmt.Errorf("Section specification error: Cannot have none and any other option")
-	}
-
-	return nil
-}
-
 func definitionFromSpec(spec string) (def types.Definition, err error) {
 
 	// Try spec as URI first
@@ -248,28 +227,5 @@ func handleRemoteBuildFlags(cmd *cobra.Command) {
 			sylog.Fatalf("Unable to get library service URI: %v", err)
 		}
 		libraryURL = uri
-	}
-}
-
-// standard builds should just warn and fall back to CLI default if we cannot resolve library URL
-func handleBuildFlags(cmd *cobra.Command) {
-	// if we can load config and if default endpoint is set, use that
-	// otherwise fall back on regular authtoken and URI behavior
-	e, err := sylabsRemote(remoteConfig)
-	if err == scs.ErrNoDefault {
-		sylog.Warningf("No default remote in use, falling back to %v", libraryURL)
-		return
-	} else if err != nil {
-		sylog.Fatalf("Unable to load remote configuration: %v", err)
-	}
-
-	authToken = e.Token
-	if !cmd.Flags().Lookup("library").Changed {
-		uri, err := e.GetServiceURI("library")
-		if err == nil {
-			libraryURL = uri
-		} else if err != nil {
-			sylog.Warningf("Unable to get library service URI: %v", err)
-		}
 	}
 }
